+++ internal/consistency/diff/schema_diff.go
@@ -47,7 +47,7 @@ type SchemaDiffCmd struct {
 	clusterNodes      []map[string]any
 	database          types.Database
 	ConnectionPool    *pgxpool.Pool
-	ConcurrencyFactor int
+	ConcurrencyFactor float64
 	BlockSize         int
 	CompareUnitSize   int
 	Output            string